+++ cscore/src/main/native/cpp/PropertyContainer.h
@@ -81,7 +81,7 @@ class PropertyContainer {
     int& ndx = m_properties[name.toStringRef(nameBuf)];
     if (ndx == 0) {
       // create a new index
-      ndx = m_propertyData.size() + 1;
+      ndx = static_cast<int>(m_propertyData.size() + 1);
       m_propertyData.emplace_back(newFunc());
     } else {
       // update existing
