+++ wpiutil/src/netconsoleTee/native/cpp/main.cpp
@@ -37,8 +37,8 @@ static bool NewlineBuffer(std::string& rem, uv::Buffer& buf, size_t len,
   wpi::StringRef toCopy = str.slice(0, idx + 1);
   if (tcp) {
     // Header is 2 byte len, 1 byte type, 4 byte timestamp, 2 byte sequence num
-    uint32_t ts = wpi::FloatToBits((wpi::Now() - startTime) * 1.0e-6);
-    uint16_t len = rem.size() + toCopy.size() + 1 + 4 + 2;
+    uint32_t ts = wpi::FloatToBits(static_cast<float>((wpi::Now() - startTime) * 1.0e-6));
+    uint16_t len = static_cast<uint16_t>(rem.size() + toCopy.size() + 1 + 4 + 2);
     out << wpi::ArrayRef<uint8_t>({static_cast<uint8_t>((len >> 8) & 0xff),
                                    static_cast<uint8_t>(len & 0xff), 12,
                                    static_cast<uint8_t>((ts >> 24) & 0xff),
